+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+### Added
+
+- Added `RoundFloatsTo` to ExecuteDatasetExtractionFlatFileDestination
+
 ### Fixed
 
 - Fixed [Pipeline] objects showing an ID of 0 in tree collections